New functionalities and RAM layout
Added function for finding files based on their path. Syscalls wrapper
function updated and now returns an integer. RAM is layed out so stack
is growing to lesser addresses from the top of the on-board RAM while
the sections for not initialized, initialized variables and memory
assigned on the heap is located in the external RAM. This insures that
the heap and the stack never collide.

Also removed test program call from mshell as it's no longer needed.

File create function is from now on depricated because in this memory
layout and with the Harvard arhitecture it makes no sense.